Convert useFocusOnBlur to TS
